handle error

+++ src/requests/Auth/disconnect.ts
@@ -6,19 +6,21 @@ import type { UserConnected } from '@/types/User'
 
 const disconnect = async () => {
 
-	await useAxios(
+	const	{ error } = await useAxios(
 		'delete',
 		'/auth/disconnect'
 	)
-	localStorage.clear()
-	const	userStore = useUserStore()
-	const	sbStore = useSideBarStore()
-	userStore.me = {} as UserConnected
-	userStore.loginApi = false
-	userStore.socket.disconnect()
-	sbStore.resetState()
-	if (!router.currentRoute.value.fullPath.includes('/accounts'))
-		router.push({ path: '/accounts' })
+	if (!error.value) {
+		localStorage.clear()
+		const	userStore = useUserStore()
+		const	sbStore = useSideBarStore()
+		userStore.me = {} as UserConnected
+		userStore.loginApi = false
+		userStore.socket.disconnect()
+		sbStore.resetState()
+		if (!router.currentRoute.value.fullPath.includes('/accounts'))
+			router.push({ path: '/accounts' })
+	}
 }
 
 export default disconnect
